Move defineEndpoints to utils package
