+++ packages/constants/package.json
@@ -14,6 +14,8 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@trpfrog.net/config-typescript": "workspace:*"
+    "@trpfrog.net/config-typescript": "workspace:*",
+    "@trpfrog.net/config-eslint": "workspace:*",
+    "@trpfrog.net/utils": "workspace:*"
   }
 }
