+++ packages/utils/src/defineEndpoints.ts
