+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 export { createURL } from './url'
 export { composeFunctions, pipeFunctions } from './composeFn'
 export { withTimeout } from './timeout'
+export { defineEndpoints } from './defineEndpoints'
